Remove item class, finished, needs notations

+++ CS1181/Lab3/Product.java
@@ -2,7 +2,7 @@
 
 public interface Product {
     
-    int getTotalPrice();
+    double getTotalPrice();
     boolean isTaxable();
 
 }
